Moves Note related Vuex code into dedicated module for the sakes of organization
